refactor(strata-storage) move NodeStorage to own crate
